+++ source/Lib/CommonLib/InterpolationFilter.cpp
@@ -791,9 +791,9 @@ void InterpolationFilter::xWeightedGeoBlk(const PredictionUnit &pu, const uint32
   Pel*    dst = predDst.get(compIdx).buf;
   Pel*    src0 = predSrc0.get(compIdx).buf;
   Pel*    src1 = predSrc1.get(compIdx).buf;
-  int32_t strideDst = predDst.get(compIdx).stride - width;
-  int32_t strideSrc0 = predSrc0.get(compIdx).stride - width;
-  int32_t strideSrc1 = predSrc1.get(compIdx).stride - width;
+  ptrdiff_t strideDst  = predDst.get(compIdx).stride - width;
+  ptrdiff_t strideSrc0 = predSrc0.get(compIdx).stride - width;
+  ptrdiff_t strideSrc1 = predSrc1.get(compIdx).stride - width;
 
   const char    log2WeightBase = 3;
   const ClpRng  clipRng = pu.cu->slice->clpRngs().comp[compIdx];
