+++ source/Lib/EncoderLib/AQp.cpp
@@ -71,7 +71,7 @@ void AQpPreanalyzer::preanalyze( Picture* pcEPic )
   const CPelBuf lumaPlane = pcEPic->getOrigBuf().Y();
   const int     width     = lumaPlane.width;
   const int     height    = lumaPlane.height;
-  const int     stride    = lumaPlane.stride;
+  const ptrdiff_t stride    = lumaPlane.stride;
 
   for ( uint32_t d = 0; d < pcEPic->aqlayer.size(); d++ )
   {
